+++ src/core/prelude/object/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.32.0 (2021-02-05)
+
+#### :rocket: New Feature
+
+* Added a new parameter `setter` for `set`
+
 ## v3.29.0 (2020-12-22)
 
 #### :rocket: New Feature